+++ src/lib/github/getStatusCommit.ts
@@ -1,5 +1,8 @@
 import { PrStatusesFragment } from "generated/graphql";
 
+/**
+ * Extract the status commit from a pull request as fetched from graphql
+ */
 export function getStatusCommit(pr: PrStatusesFragment) {
-  return pr.commits.nodes?.[0]?.commit;
+  return pr.commits?.nodes?.[0]?.commit;
 }
